requirements.txt
streamlit
qrcode[pil]
opencv-python
pandas
Pillow
numpy
